deans-api/requirements.txt
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
lazy-object-proxy==1.3.1
Markdown==2.6.11
pytz==2018.5
psycopg2-binary
gunicorn==19.9.0
django-cron==0.5.1
django-cors-headers==2.4.0
gevent==1.3.7
django-rest-auth==0.9.3
django-extensions==2.1.3
requests==2.20.0
channels==2.1.5
channels_redis==2.3.1
Pygments==2.2.0
Markdown==2.6.11
coreapi==2.3.3
docutils==0.14
